new object file code
reads the object file into a separate array instead of just parsing it in memory..
checks for invalid file size and if offsets are too large.
if offsets are umm. off.. it decrypts it again to see if it will work.
supports amiga 4 byte object items (untested)
may change bits 'n pieces later on..