new dir preset system
instead of having a bunch of dir's everywhere.. it now uses a central preset array of dirs.. it's abstracted enough so i can eventually optimise it so the data shares common pathes and saves memory.

basically presets like game, nagi, user, save and other dir's are set when they're available and nagi calls the preset system to change to these directories when it needs to be guaranteed that it's at that certain directory.

added..  new sys_dir.c preset system
hooks in various sources to change dir at a certain time.